voice_integration/requirements.txt
firecrawl>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
flask>=2.0.0
quart>=0.18.0
aiohttp>=3.8.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
hypercorn>=0.14.4
elevenlabs>=0.2.24
pyaudio>=0.2.13
numpy>=1.24.0
websockets>=11.0.3
